backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Gateway for Amazon Bedrock APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "uvicorn>=0.41.0",
    "boto3>=1.42.58",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "gunicorn>=25.1.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-botocore>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-instrumentation-urllib3>=0.57b0",
    "opentelemetry-instrumentation-system-metrics>=0.57b0",
    "opentelemetry-instrumentation>=0.57b0",
    "pyjwt[crypto]>=2.11.0",
    "requests>=2.32.5",
    "jsonschema>=4.26.0",
    "valkey-glide>=2.2.7",
    "pyyaml>=6.0.3",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "README.md"
]
